refactor: split create device form handling
